13/05 ליאל

+++ ClassLibrary1/AssistedBL.cs
@@ -29,33 +29,35 @@ namespace BL
                 assisted a = db.assisted.FirstOrDefault(x => x.id_assisted == assisted.id_assisted);
                 if (a != null)
                 {
-                    List<assisted_language> lList = new List<assisted_language>();
-                    List<assisted_language> alList = db.assisted_language.Where(x => x.id_assisted == assisted.id_assisted).ToList();
-                    foreach (Language item in assisted.languages)
+                    if (assisted.languages != null)
                     {
-                        if (item.IsSelected == true)
+                        List<assisted_language> lList = new List<assisted_language>();
+                        List<assisted_language> alList = db.assisted_language.Where(x => x.id_assisted == assisted.id_assisted).ToList();
+                        foreach (Language item in assisted.languages)
                         {
-                            assisted_language temp = alList.FirstOrDefault(x => x.id_assisted == assisted.id_assisted && x.code_language == item.code_language);
-                            if (temp == null)
+                            if (item.IsSelected == true)
                             {
-                                assisted_language al = new assisted_language();
-                                al.id_assisted = assisted.id_assisted;
-                                al.code_language = item.code_language;
-                                lList.Add(al);
+                                assisted_language temp = alList.FirstOrDefault(x => x.id_assisted == assisted.id_assisted && x.code_language == item.code_language);
+                                if (temp == null)
+                                {
+                                    assisted_language al = new assisted_language();
+                                    al.id_assisted = assisted.id_assisted;
+                                    al.code_language = item.code_language;
+                                    lList.Add(al);
+                                }
+                                else
+                                    alList.Remove(temp);
                             }
-                            else
-                                alList.Remove(temp);
+                        }
+                        foreach (assisted_language item in lList)
+                        {
+                            db.assisted_language.Add(item);
+                        }
+                        foreach (assisted_language item in alList)
+                        {
+                            db.assisted_language.Remove(item);
                         }
                     }
-                    foreach (assisted_language item in lList)
-                    {
-                        db.assisted_language.Add(item);
-                    }
-                    foreach (assisted_language item in alList)
-                    {
-                        db.assisted_language.Remove(item);
-                    }
-
 
                     List<assisted_domain> dList = new List<assisted_domain>();
                     List<assisted_domain> adList = db.assisted_domain.Where(x => x.id_assisted == assisted.id_assisted).ToList();
@@ -68,7 +70,6 @@ namespace BL
                             {
                                 assisted_domain ad = new assisted_domain();
                                 ad.id_assisted = assisted.id_assisted;
-                                //vvo.code_volunteering = db.volunteering_domain.Where(x => x.description == item.description).Select(s => s.code_volunteering).FirstOrDefault();
                                 ad.code_volunteering = item.code_volunteering;
                                 dList.Add(ad);
                             }
@@ -84,40 +85,35 @@ namespace BL
                             db.assisted_domain.Remove(item);
                         }
                     }
-                    var aval = db.availability.Where(x => x.code_day == assisted.availability.code_day && x.code_shift == assisted.availability.code_shift).FirstOrDefault();
-                    assisted_availability availability = new assisted_availability();
-                    availability.code_availability = aval.code_availability;
-                    availability.id_assisted = assisted.id_assisted;
-                    assisted_availability a_v = db.assisted_availability.Where(x => x.id_assisted == availability.id_assisted).FirstOrDefault();
-                    a_v.code_availability = availability.code_availability;
-                    List<assisted_availability> aList = new List<assisted_availability>();
-                    List<assisted_availability> aaList = db.assisted_availability.Where(x => x.id_assisted == assisted.id_assisted).ToList();
-                    foreach (Availability item in assisted.availabilitys)
+                    if (assisted.availabilitys != null)
                     {
-                        if (item.IsSelected == true)
+                        List<assisted_availability> aList = new List<assisted_availability>();
+                        List<assisted_availability> aaList = db.assisted_availability.Where(x => x.id_assisted == assisted.id_assisted).ToList();
+                        foreach (Availability item in assisted.availabilitys)
                         {
-                            assisted_availability temp = aaList.FirstOrDefault(x => x.id_assisted == assisted.id_assisted && x.code_availability == item.code_availability);
-                            if (temp == null)
+                            if (item.IsSelected == true)
                             {
-                                assisted_availability av = new assisted_availability();
-                                av.id_assisted = assisted.id_assisted;
-                                av.code_availability = item.code_availability;
-                                aList.Add(av);
+                                assisted_availability temp = aaList.FirstOrDefault(x => x.id_assisted == assisted.id_assisted && x.code_availability == item.code_availability);
+                                if (temp == null)
+                                {
+                                    assisted_availability av = new assisted_availability();
+                                    av.id_assisted = assisted.id_assisted;
+                                    av.code_availability = item.code_availability;
+                                    aList.Add(av);
+                                }
+                                else
+                                    aaList.Remove(temp);
                             }
-                            else
-                                aaList.Remove(temp);
+                        }
+                        foreach (assisted_availability item in aList)
+                        {
+                            db.assisted_availability.Add(item);
+                        }
+                        foreach (assisted_availability item in aaList)
+                        {
+                            db.assisted_availability.Remove(item);
                         }
                     }
-                    foreach (assisted_availability item in aList)
-                    {
-                        db.assisted_availability.Add(item);
-                    }
-                    foreach (assisted_availability item in aaList)
-                    {
-                        db.assisted_availability.Remove(item);
-                    }
-
-
                     assisted a1 = Assisted.convertassistedentitytoassistedtable(assisted);      
                     a.id_assisted = a1.id_assisted;
                     a.first_name = a1.first_name;
@@ -140,75 +136,36 @@ namespace BL
                 }
                 else
                 {
-                    assisted a1 = Assisted.convertassistedentitytoassistedtable(assisted);
-                    assisted ass = new assisted();
-                    ass.id_assisted = a1.id_assisted;
-                    ass.first_name = a1.first_name;
-                    ass.last_name = a1.last_name;
-                    ass.date_birth = a1.date_birth;
-                    ass.code_gender = a1.code_gender;
-                    ass.code_city = a1.code_city;
-                    ass.code_status = a1.code_status;
-                    ass.e_mail = a1.e_mail;
-                    ass.first_name = a1.first_name;
-                    ass.id_assisted = a1.id_assisted;
-                    ass.last_name = a1.last_name;
-                    ass.street = a1.street;
-                    ass.number_building = a1.number_building;
-                    ass.number_floor = a1.number_floor;
-                    ass.number_house = a1.number_house;
-                    ass.password = a1.password;
-                    ass.phone = a1.phone;
-                    ass.postal_code = a1.postal_code;
                     db.assisted.Add(Assisted.convertassistedentitytoassistedtable(assisted));
-                    //db.SaveChanges();
-
-
+                    db.SaveChanges();
 
                     List<assisted_language> alList = Assisted.ConvertLanguageEntityListToAssistedLanguage(assisted.languages, assisted.id_assisted);
                     foreach (assisted_language l in alList)
                     {
                         db.assisted_language.Add(l);
                     }
-                    db.assisted.Add(Assisted.convertassistedentitytoassistedtable(assisted));
 
                     List<assisted_availability> AvailabilityList = Assisted.ConvertAvailabilityEntityListToAssistedAvailability(assisted.availabilitys, assisted.id_assisted);
                     foreach (assisted_availability al in AvailabilityList)
                     {
                         db.assisted_availability.Add(al);
                     }
-                    db.assisted.Add(Assisted.convertassistedentitytoassistedtable(assisted));
 
-                    List<assisted_domain> dList = new List<assisted_domain>();
-                    foreach (VolunteeringDomain item in assisted.volunteeringdomains)
+                    if (assisted.volunteeringdomains != null)
                     {
-                        assisted_domain ad = new assisted_domain();
-                        ad.id_assisted = assisted.id_assisted;
-                        ad.code_volunteering = item.code_volunteering;
-                        dList.Add(ad);
+                        List<assisted_domain> dList = new List<assisted_domain>();
+                        foreach (VolunteeringDomain item in assisted.volunteeringdomains)
+                        {
+                            assisted_domain ad = new assisted_domain();
+                            ad.id_assisted = assisted.id_assisted;
+                            ad.code_volunteering = item.code_volunteering;
+                            dList.Add(ad);
+                        }
+                        foreach (assisted_domain item in dList)
+                        {
+                            db.assisted_domain.Add(item);
+                        }
                     }
-                    foreach (assisted_domain item in dList)
-                    {
-                        db.assisted_domain.Add(item);
-                    }
-                    db.assisted.Add(Assisted.convertassistedentitytoassistedtable(assisted));
-                    //List<assisted_domain> dList = new List<assisted_domain>();
-                    //foreach (string item in assisted.domain)
-                    //{
-                    //    assisted_domain vvo = new assisted_domain();
-                    //    vvo.id_assisted = assisted.id_assisted;
-                    //    vvo.code_volunteering = db.volunteering_domain.Where(x => x.description == item).Select(s => s.code_volunteering).FirstOrDefault();
-                    //    dList.Add(vvo);
-                    //}
-                    //foreach (assisted_domain item in dList)
-                    //{
-                    //    db.assisted_domain.Add(item);
-                    //}
-                    //var aval = db.availability.Where(x => x.code_day == assisted.availability.code_day && x.code_shift == assisted.availability.code_shift).FirstOrDefault();
-                    //assisted_availability availability = new assisted_availability();
-                    //availability.code_availability = aval.code_availability;
-                    //availability.id_assisted = assisted.id_assisted;
-                    //db.assisted_availability.Add(availability);
                 }
 
                 db.SaveChanges();
@@ -220,45 +177,54 @@ namespace BL
 
             return true;
         }
-        public static bool EmbededAssisted(Assisted assisted)
+        //פונקציית שיבוץ
+        public static List<Assisted> EmbededAssisteds(Volunteer volunteer)
         {
-            List<volunteer> vDomain = new List<volunteer>();
-            var domainAssisted = db.assisted_domain.Where(ad => ad.id_assisted == assisted.id_assisted).ToList();
-            foreach (var item in db.volunteer.ToList())
+            List<assisted> assisteds = new List<assisted>();
+            var domainVolunteer = db.volunteer_domain.Where(vd => vd.id_volunteer == volunteer.id_volunteer).ToList();
+            foreach (var item in db.assisted.ToList())
             {
-                foreach (var vd in item.volunteer_domain)
+                foreach (var ad in item.assisted_domain)
                 {
-                    if(domainAssisted.Find(x=>x.code_volunteering == vd.code_volunteering) != null)
+                    if (domainVolunteer.Find(x => x.code_volunteering == ad.code_volunteering) != null)
                     {
-                        if (!vDomain.Contains(item))
+                        if (!assisteds.Contains(item))
                         {
-                            vDomain.Add(item);
+                            assisteds.Add(item);
                         }
                     }
                 }
             }
-            var avalAssisted = db.assisted_availability.Where(av => av.id_assisted == assisted.id_assisted).ToList();
-            foreach (var item in vDomain)
+            var avalVolunteer = db.availability_volunteer.Where(av => av.id_volunteer == volunteer.id_volunteer).ToList();
+            foreach (var item in assisteds)
             {
-                var list = db.availability_volunteer.Where(x => x.id_volunteer == item.id_volunteer).ToList();
-                for (int i = 0; i < avalAssisted.Count; i++)
+                var avalAssisted = db.assisted_availability.Where(x => x.id_assisted == item.id_assisted).ToList();
+                for (int i = 0; i < avalVolunteer.Count; i++)
                 {
-                    if(list.Find(x=>x.code_availability == avalAssisted[i].code_availability) != null)
+                    if (avalAssisted.Find(x => x.code_availability == avalVolunteer[i].code_availability) == null)
                     {
-
-                        return true;
+                        assisteds.Remove(item);
                     }
                 }
             }
-            return false;
-            //var domaim = db.volunteer_domain
-            //db.volunteer.Where(v=>v.code_volunteering_domain == )
+            var langVolunteer = db.volunteer_language.Where(vl => vl.id_volunteer == volunteer.id_volunteer).ToList();
+            foreach (var item in assisteds)
+            {
+                var langAssisted = db.assisted_language.Where(x => x.id_assisted == item.id_assisted).ToList();
+                for (int i = 0; i < langVolunteer.Count; i++)
+                {
+                    if (langAssisted.Find(x => x.code_language == langVolunteer[i].code_language) == null)
+                    {
+                        assisteds.Remove(item);
+                    }
+                }
+            }
+            assisteds = assisteds.Where(a => a.code_city == volunteer.code_city).ToList();
 
-            //SendMail(assisted.e_mail, "ddd" , "ddd");
+            return Assisted.convertassistedtabletolistassistedentity(assisteds);
 
-
+            //SendMail(assisted.e_mail, "ddd", "ddd");
         }
-
         public static List<Assisted> RemoveAssisted(string id_assisted)
         {
             db.assisted.Remove(db.assisted.FirstOrDefault(x => x.id_assisted == id_assisted));
